+запросы подтверждения

+++ telephone_system/src/main/webapp/theme1/static/scripts/tableDepartment.js
@@ -82,7 +82,7 @@ function loadADSLTableDel(elem){
     			'</thead>';
     			for(var i=0;i < parseInt(data.name.length); i++){
     					adslList += '<tbody>'+
-    									'<tr><td>' + (i+1) + '</td><td class="info">'+data.name[i]+'</td><td width="20px">      <button id = "'+data.name[i]+'" class="del" style="cursor:pointer" onClick = "getdetails(this)"><img src="styles/kartoteka/img/tableDel.png" style="vertical-align: middle"></img></button>        </td></tr>'+
+    									'<tr><td width="20px">' + (i+1) + '</td><td class="info">'+data.name[i]+'</td><td width="20px">      <button id = "'+data.name[i]+'" class="del" style="cursor:pointer" onClick = "getdetails(this)"><img src="styles/kartoteka/img/tableDel.png" style="vertical-align: middle"></img></button>        </td></tr>'+
     								'</tbody>';
     			}
     			adslList+='</table>';
@@ -201,28 +201,30 @@ $(document).ready(function() {
 
 //Обработка нажатия кнопки удалеия
 function getdetails(obj) {
-	var param = obj.id;
-	var param_ = [];
-	param_.push(param);
-	if(param!=''){
-		$.ajax({
-			type: 'DELETE',
-			url:  mainURL + '/departmentList',
-			contentType: 'application/json; charset=utf-8',
-			data: JSON.stringify(param_),
-			dataType: 'json',
-			async: true,
-			success: function(result) {
-				alert('Статус: ' + result);
-				loadADSLTable(1);
-			},
-			error: function(jqXHR, textStatus, errorThrown) {
-				alert('Статус: ' + jqXHR.responseText);
-				loadADSLTable(1);
-			}
-		});
-		blockInput();
-		loadADSLTable(page);
+	if (confirm("Вы точно хотите удалить подразделение?")) {
+		var param = obj.id;
+		var param_ = [];
+		param_.push(param);
+		if(param!=''){
+			$.ajax({
+				type: 'DELETE',
+				url:  mainURL + '/departmentList',
+				contentType: 'application/json; charset=utf-8',
+				data: JSON.stringify(param_),
+				dataType: 'json',
+				async: true,
+				success: function(result) {
+					alert('Статус: ' + result);
+					loadADSLTable(1);
+				},
+				error: function(jqXHR, textStatus, errorThrown) {
+					alert('Статус: ' + jqXHR.responseText);
+					loadADSLTable(1);
+				}
+			});
+			blockInput();
+			loadADSLTable(page);
+		}
 	}
 }
